Extracting generic service logic into base service module
